xen: move Xen hypercall wrapper functions to sys/xen
There is minimal variation between architectures for the hypercall
wrappers.  Most common is word size, which effects a few wrappers, for
these check for ILP32 (which identifies all 32-bit architectures).  Also
notable is HYPERVISOR_set_trap_table() which appears x86-only.  Due to
the minimal difference, move most of the upper level wrappers to a
common header, leaving only the lower level wrappers with the
architecture.

SPDX tags were added.

Add comment documenting _hypercall#() interface with an eye towards
future development.

The amd64 definition of HYPERVISOR_set_callbacks() was wrong and the
i386 was used instead (last argument is ignored by Xen on amd64, but the
definition was still wrong).