kern/intr: add intr_event as "source" value
Callers of `intr_event_init()` are expected to convert to passing a
structure surrounding the event structure.  As a result the pointers
alias and a pointer to the event can be converted back to the "source".

This makes the relationship between machine interrupt core and interrupt
events, similar to the relationship between PICs and machine interrupt
core.

As `intr_event_init()` does not sleep, it can be called while locks are
held.  This allows cleanup of `intr_register_source()` as the table
pre-check is now worthless.