Update rc.initdiskless, fix error handling of remount_optional
chkerr() ignores the exit code of a preceding mount command in case
a file ```remount_optional``` exists. The check is performed and a
subshell is launched to log the informational message and return.
The return is executed in the context of the subshell, not the
context of the chkerr() function, hence is a NOP.  The
remount_optional check is hence ineffective.

Change the code to if/then/fi, so the return is evaluated in the
context of the  chkerr function, to make the check effective.